requirements.txt
toml
plotly
streamlit
mysql-connector-python
pandas
